refactored function locations and piece component locations
