+++ src/lib/components/game/moves/MoveBox.svelte
@@ -1,6 +1,6 @@
 <script>
-	import { getPiececomponent } from '../../../../functions/getPieceComponent';
-	import { moves } from '../../../../stores/moves';
+	import { getPiececomponent } from '../../../../functions/rendering/getPieceComponent';
+	import { moves } from '../../../stores/moves';
 	import Spinner from './Spinner.svelte';
 </script>
