+++ src/lib/components/game/pieces/Bishop_B.svelte
