+++ src/lib/components/game/pieces/Bishop_W.svelte
