+++ src/lib/components/game/pieces/Horse_B.svelte
