+++ src/lib/components/game/pieces/Horse_W.svelte
