+++ src/lib/components/game/pieces/King_B.svelte
