+++ src/lib/components/game/pieces/King_W.svelte
