+++ src/lib/components/game/pieces/Pawn_B.svelte
