+++ src/lib/components/game/pieces/Pawn_W.svelte
