+++ src/lib/components/game/pieces/Queen_B.svelte
