+++ src/lib/components/game/pieces/Queen_W.svelte
