+++ src/lib/components/game/pieces/Tower_B.svelte
