+++ src/lib/components/game/pieces/Tower_W.svelte
