+++ src/lib/functions/game/moves/checkedMoves.ts
@@ -1,0 +1,23 @@
+import type { Piece, Square } from '../../../stores/stores';
+import { kingChecked } from '../../checked/kingChecked';
+
+export const moveAllowedWhileCheck = (
+	board: Square[],
+	newLoc: number,
+	currentLoc: number,
+	kingLoc: number,
+	movingPiece: Piece,
+	turn: 'black' | 'white'
+) => {
+	let testBoard = JSON.parse(JSON.stringify(board));
+	testBoard[currentLoc].piece = null;
+	testBoard[newLoc].piece = { type: movingPiece.type, color: movingPiece.color };
+
+	if (movingPiece.type == 'king') {
+		if (kingChecked(testBoard, { type: 'king', color: turn }, newLoc)) return false;
+		else return true;
+	} else {
+		if (kingChecked(testBoard, { type: 'king', color: turn }, kingLoc)) return false;
+		else return true;
+	}
+};
