+++ src/lib/functions/rendering/board.ts
@@ -1,0 +1,5 @@
+export const letters = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H'] as const;
+export const darkSquares = [
+	1, 3, 5, 7, 8, 10, 12, 14, 17, 19, 21, 23, 24, 26, 28, 30, 33, 35, 37, 39, 40, 42, 44, 46, 49, 51,
+	53, 55, 56, 58, 60, 62, 64
+];
