+++ src/routes/+page.svelte
@@ -10,7 +10,7 @@
 		browserSessionPersistence
 	} from 'firebase/auth';
 	import { doc, getDoc, setDoc } from 'firebase/firestore';
-	import { userStore } from '../stores.js';
+	import { userStore } from '../lib/stores/stores.js';
 	import ForgotPw from '$lib/components/signin/ForgotPw.svelte';
 
 	let createAccount = false;
