+++ src/routes/game/[[gameId]]/[[userId]]/+page.svelte
@@ -6,7 +6,7 @@
 	import Resigned from '$lib/components/game/modals/Resigned.svelte';
 	import GameEnded from '$lib/components/game/modals/GameEnded.svelte';
 	import Opponents from '$lib/components/game/players/Opponents.svelte';
-	import { gameId, player, resign, url, userId, winner } from '../../../../stores';
+	import { gameId, player, resign, url, userId, winner } from '../../../../lib/stores/stores';
 	import type { PageData } from './$types';
 	import { page } from '$app/stores';
 	import { onMount } from 'svelte';
